fix bad merge; proper test split
