+++ java/core/src/test/java/co/worklytics/psoxy/impl/RESTApiSanitizerImplTest.java
@@ -65,8 +65,6 @@ class RESTApiSanitizerImplTest {
     @Inject
     protected PseudonymizerImplFactory pseudonymizerImplFactory;
 
-    ObjectMapper objectMapper = new ObjectMapper();
-
     @Inject
     ConfigService config;
     @Inject
@@ -75,7 +73,10 @@ class RESTApiSanitizerImplTest {
 
     @Singleton
     @Component(
-            modules = {PsoxyModule.class, ForConfigService.class, MockModules.ForSecretStore.class,
+            modules = {
+                PsoxyModule.class,
+                ForConfigService.class,
+                MockModules.ForSecretStore.class,
             // TestModules.ForSecretStore.class,
             })
     public interface Container {
@@ -203,77 +204,8 @@ class RESTApiSanitizerImplTest {
         assertFalse(sanitizer.isAllowed("GET", new URL(url)));
     }
 
-    @SneakyThrows
-    @ValueSource(strings = {"pwd=1234asAf", " pwd=1234asAf  ",
-            "https://asdf.google.com/asdf/?pwd=1234asAf",
-            "https://asdf.google.com/asdf/?pwd=1234asAf&pwd=14324",
-            "https://asdf.google.com/asdf/?asdf=2134&pwd=1234asAf&",
-            "https://asdf.google.com/asdf/?asdf=2134&PWD=1234asAf&",
-            "https://asdf.google.com/asdf/?asdf=2134&Pwd=1234asAf&"})
-    @ParameterizedTest
-    void redactRegexMatches(String source) {
-        Transform.RedactRegexMatches transform =
-                Transform.RedactRegexMatches.builder().redaction("(?i)pwd=[^&]*").build();
-
-        assertTrue(StringUtils.containsIgnoreCase(source, "pwd=1234asAf"));
-        String redacted = (String) sanitizer.sanitizerUtils.getRedactRegexMatches(transform)
-                .map(source, sanitizer.jsonConfiguration);
-        assertFalse(StringUtils.containsIgnoreCase(redacted, "pwd=1234asAf"));
-    }
 
 
-    @SneakyThrows
-    @CsvSource({"phrase1,phrase1", "Phrase1,Phrase1", "phrase2 2,phrase2 2", "phrase3,phrase3",
-            "phrase4,", "blah phrase1,phrase1", "blah phrase1: blah,phrase1",})
-    @ParameterizedTest
-    void redactExceptPhases(String raw, String sanitized) {
-        Transform.RedactExceptPhrases transform = Transform.RedactExceptPhrases.builder()
-                .allowedPhrases(Arrays.asList("phrase1", "phrase2 2", "Phrase3")).build();
-
-        String redacted = (String) sanitizer.sanitizerUtils.getRedactExceptPhrases(transform)
-                .map(raw, sanitizer.jsonConfiguration);
-
-        assertEquals(sanitized, StringUtils.trimToNull(redacted));
-    }
-
-    @SneakyThrows
-    @ValueSource(strings = {"https://acme.zoom.us/12312345?pwd=1234asAf asdfasdf",
-            " https://acme.zoom.us/12312345?pwd=1234asAf  ",
-            "come to my zoom meeting! https://acme.zoom.us/12312345?pwd=1234asAf",
-            "come to my zoom meeting! \r\n https://acme.zoom.us/12312345?pwd=1234asAf",
-            "come to my zoom meeting! \r\nhttps://acme.zoom.us/12312345?pwd=1234asAf\r\n",
-            "come to my zoom meeting! \r\n this is the url: https://acme.zoom.us/12312345?pwd=1234asAf",
-            "come to my zoom meeting! \r\n this is the url: https://acme.zoom.us/12312345?pwd=1234asAf\r\nthat was the url",
-            "https://acme.zoom.us/12312345?pwd=1234asAf"})
-    @ParameterizedTest
-    void filterTokensByRegex(String source) {
-        Transform.FilterTokenByRegex transform = Transform.FilterTokenByRegex.builder()
-                .filter("https://[^.]+\\.zoom\\.us/.*").build();
-
-        String redacted = (String) sanitizer.sanitizerUtils.getFilterTokenByRegex(transform)
-                .map(source, sanitizer.jsonConfiguration);
-
-        assertEquals("https://acme.zoom.us/12312345?pwd=1234asAf", redacted);
-    }
-
-    @SneakyThrows
-    @ValueSource(strings = {"", " https://acme.meet.us/12312345?pwd=1234asAf  ",
-            "come to my zoom meeting! https://acme.meet.us/12312345?pwd=1234asAf",
-            "come to my zoom meeting! \r\n https://acme.meet.us/12312345?pwd=1234asAf",
-            "come to my zoom meeting! \r\nhttps://zoom.us/12312345?pwd=1234asAf\r\n",
-            "come to my zoom meeting! \r\n this is the url: https://acme.asdfasd.zoom.us/12312345?pwd=1234asAf",
-            "come to my zoom meeting! \r\n this is the url: https://acme/zoom.us/12312345?pwd=1234asAf\r\nthat was the url",
-            "https://acme/zoom.us/12312345?pwd=1234asAf", "  ", "\r\n",})
-    @ParameterizedTest
-    void filterTokensByRegex_rejects(String source) {
-        Transform.FilterTokenByRegex transform = Transform.FilterTokenByRegex.builder()
-                .filter("https://[^.]+\\.zoom\\.us/.*").build();
-
-        String redacted = (String) sanitizer.sanitizerUtils.getFilterTokenByRegex(transform)
-                .map(source, sanitizer.jsonConfiguration);
-
-        assertTrue(StringUtils.isBlank(redacted));
-    }
 
     @SneakyThrows
     @ValueSource(strings = {"\"https://asdf.google.com/asdf/?asdf=2134&Pwd=1234asAf&\"", // url as
@@ -326,40 +258,6 @@ class RESTApiSanitizerImplTest {
     }
 
 
-    @Test
-    void pseudonymizeWithReversalKey() {
-        // NOTE: this is a LEGACY case
-        MapFunction f = sanitizer.sanitizerUtils.getPseudonymize(sanitizer.getPseudonymizer(),
-                Transform.Pseudonymize.builder().includeReversible(true).build());
-
-        assertEquals(
-                "{\"hash\":\"Z7Bnl_VVOwSmfP9kuT0_Ub-5ic4cCVI4wCHArL1hU0M\",\"reversible\":\"p~Z7Bnl_VVOwSmfP9kuT0_Ub-5ic4cCVI4wCHArL1hU0MzTTbTCc7BcR53imT1qZgI\"}",
-                f.map("asfa", sanitizer.getJsonConfiguration()));
-    }
-
-    @Test
-    void reversiblePseudonym() {
-        MapFunction f = sanitizer.sanitizerUtils.getPseudonymize(sanitizer.getPseudonymizer(),
-                Transform.Pseudonymize.builder().includeReversible(true).build());
-
-        String lcase = (String) f.map("erik@engetc.com", sanitizer.getJsonConfiguration());
-        String ucaseFirst = (String) f.map("Erik@engetc.com", sanitizer.getJsonConfiguration());
-
-        assertNotEquals(lcase, ucaseFirst);
-        // but hashes the same
-        assertEquals(lcase.substring(0, 32), ucaseFirst.substring(0, 32));
-    }
-
-
-    @Test
-    void tokenize() {
-        String original = "blah";
-        MapFunction f = sanitizer.sanitizerUtils.getTokenize(Transform.Tokenize.builder().build());
-        String r = (String) f.map(original, sanitizer.getJsonConfiguration());
-
-        assertArrayEquals(reversibleTokenizationStrategy.getReversibleToken(original),
-                pseudonymEncoder.decode(r).getReversible());
-    }
 
     @Test
     void tokenize_regex() {
@@ -376,58 +274,8 @@ class RESTApiSanitizerImplTest {
                 reversibleTokenizationStrategy));
     }
 
-    @Test
-    void tokenize_regex_on_the_middle() {
-        String path =
-                "/v1.0/users/48d31887-5fad-4d73-a9f5-3c356e68a038/calendars('AAMkAGVmMDEzMTM4LTZmYWUtNDdkNC1hMDZiLTU1OGY5OTZhYmY4OABGAAAAAAAiQ8W967B7TKBjgx9rVEURBwAiIsqMbYjsT5e-T7KzowPTAAAAAAEGAAAiIsqMbYjsT5e-T7KzowPTAAABuC35AAA%3D')/events";
-        String host = "https://graph.microsoft.com";
-        MapFunction f = sanitizer.sanitizerUtils.getTokenize(Transform.Tokenize.builder()
-                .regex("^https://graph.microsoft.com/v1.0/users/([a-zA-Z0-9_-]+)/.*$").build());
-        String r = (String) f.map(host + path, sanitizer.getJsonConfiguration());
 
-        assertNotEquals(host + path, r);
-        assertEquals(host + path, pseudonymEncoder.decodeAndReverseAllContainedKeyedPseudonyms(r,
-                reversibleTokenizationStrategy));
-    }
 
-    @SneakyThrows
-    @ParameterizedTest
-    @CsvSource({"'Hello world', 2, 11", "'This is a test', 4, 14", "'', 0, 0", "'OneWord', 1, 7"})
-    void textDigest(String input, int expectedWordCount, int expectedLength) {
-        Transform.TextDigest transform = Transform.TextDigest.builder().build();
-        MapFunction textDigestFunction = sanitizer.sanitizerUtils.getTextDigest(transform);
-
-        String resultJson =
-                (String) textDigestFunction.map(input, sanitizer.getJsonConfiguration());
-        Map<String, Integer> result = new ObjectMapper().readValue(resultJson, Map.class);
-
-        assertEquals(expectedWordCount, result.get("word_count"));
-        assertEquals(expectedLength, result.get("length"));
-    }
-
-    @SneakyThrows
-    @Test
-    void textDigest_with_escaping() {
-        String input = "{\n" + "  \"type\": \"AdaptiveCard\",\n" + "  \"version\": \"1.0\",\n"
-                + "  \"body\": [\n" + "    {\n" + "      \"type\": \"TextBlock\",\n"
-                + "      \"text\": \"It looks like there were no important \\\"emails\\\" from last week. However, I found some relevant meetings and files that might be of interest to you.\\n\\nFrom your meetings last week:\\n- **[test meeting2 - export api](https://teams.microsoft.com/l/meeting/details?eventId=AAMkADcyZTMzNWZhLWE1YjAtNDc3Mi04MzBlLTc2NzEzOTE0MmU1ZQBGAAAAAAC5e4DRHIMCQJ-tS6nB82CZBwCMIOyf3WTwTIsBMwZamp77AAAAAAENAACMIOyf3WTwTIsBMwZamp77AABCrxI6AAA%3d)**: You discussed the need to send a reminder about an upcoming event, possibly Ignite, scheduled for next week. You emphasized the importance of the event and the reminder[1](https://teams.microsoft.com/l/meeting/details?eventId=AAMkADcyZTMzNWZhLWE1YjAtNDc3Mi04MzBlLTc2NzEzOTE0MmU1ZQBGAAAAAAC5e4DRHIMCQJ-tS6nB82CZBwCMIOyf3WTwTIsBMwZamp77AAAAAAENAACMIOyf3WTwTIsBMwZamp77AABCrxI6AAA%3d).\\n- **[new meeting to test copilot interaction in meetings](https://teams.microsoft.com/l/meeting/details?eventId=AAMkADcyZTMzNWZhLWE1YjAtNDc3Mi04MzBlLTc2NzEzOTE0MmU1ZQBGAAAAAAC5e4DRHIMCQJ-tS6nB82CZBwCMIOyf3WTwTIsBMwZamp77AAAAAAENAACMIOyf3WTwTIsBMwZamp77AABCrxI5AAA%3d)**: This meeting was held last Friday from 12:30 PM to 1 PM[2](https://teams.microsoft.com/l/meeting/details?eventId=AAMkADcyZTMzNWZhLWE1YjAtNDc3Mi04MzBlLTc2NzEzOTE0MmU1ZQBGAAAAAAC5e4DRHIMCQJ-tS6nB82CZBwCMIOyf3WTwTIsBMwZamp77AAAAAAENAACMIOyf3WTwTIsBMwZamp77AABCrxI5AAA%3d).\\n- **[teste meeting](https://teams.microsoft.com/l/meeting/details?eventId=AAMkADcyZTMzNWZhLWE1YjAtNDc3Mi04MzBlLTc2NzEzOTE0MmU1ZQBGAAAAAAC5e4DRHIMCQJ-tS6nB82CZBwCMIOyf3WTwTIsBMwZamp77AAAAAAENAACMIOyf3WTwTIsBMwZamp77AABAvsP6AAA%3d)**: You explained the significance of the Nobel Prize in Economics and announced the 2024 Nobel Prize winners, Darren Simon Johnson and James A. Robinson[3](https://teams.microsoft.com/l/meeting/details?eventId=AAMkADcyZTMzNWZhLWE1YjAtNDc3Mi04MzBlLTc2NzEzOTE0MmU1ZQBGAAAAAAC5e4DRHIMCQJ-tS6nB82CZBwCMIOyf3WTwTIsBMwZamp77AAAAAAENAACMIOyf3WTwTIsBMwZamp77AABAvsP6AAA%3d).\\n\\nAdditionally, there is a file titled **[OnCall DRI Handbook-v3](https://m365cpi17278319-my.sharepoint.com/personal/corat_m365cpi17278319_onmicrosoft_com/Documents/Microsoft%20Copilot%20Chat%20Files/OnCall%20DRI%20Handbook-v3.pdf?web=1)** that you last modified on February 4th, 2021. This document provides guidelines on handling incidents and includes important terminology and procedures[4](https://m365cpi17278319-my.sharepoint.com/personal/corat_m365cpi17278319_onmicrosoft_com/Documents/Microsoft%20Copilot%20Chat%20Files/OnCall%20DRI%20Handbook-v3.pdf?web=1).\\n\\nIs there anything specific you would like to know more about?\",\n"
-                + "      \"wrap\": true\n" + "    },\n" + "    {\n"
-                + "      \"type\": \"TextBlock\",\n" + "      \"id\": \"MessageTextField\",\n"
-                + "      \"text\": \"It looks like there were no important \\\"emails\\\" from last week. However, I found some relevant meetings and files that might be of interest to you.\\n\\nFrom your meetings last week:\\n- **test meeting2 - export api[3]**: You discussed the need to send a reminder about an upcoming event, possibly Ignite, scheduled for next week. You emphasized the importance of the event and the reminder[^2^].\\n- **new meeting to test copilot interaction in meetings[3]**: This meeting was held last Friday from 12:30 PM to 1 PM[^3^].\\n- **teste meeting[3]**: You explained the significance of the Nobel Prize in Economics and announced the 2024 Nobel Prize winners, Darren Simon Johnson and James A. Robinson[^4^].\\n\\nAdditionally, there is a file titled **OnCall DRI Handbook-v3[2]** that you last modified on February 4th, 2021. This document provides guidelines on handling incidents and includes important terminology and procedures[^1^].\\n\\nIs there anything specific you would like to know more about?\",\n"
-                + "      \"wrap\": true\n" + "    }\n" + "  ]\n" + "}";
-
-        String expected =
-                "{\"type\":\"AdaptiveCard\",\"version\":\"1.0\",\"body\":[{\"type\":\"TextBlock\",\"text\":\"{\\\"length\\\":982,\\\"word_count\\\":154}\",\"wrap\":true},{\"type\":\"TextBlock\",\"id\":\"MessageTextField\",\"text\":\"{\\\"length\\\":982,\\\"word_count\\\":154}\",\"wrap\":true}]}";
-
-        Transform.TextDigest transform = Transform.TextDigest.builder().isJsonEscaped(true)
-                .jsonPathToProcessWhenEscaped("$..text").build();
-        MapFunction textDigestFunction = sanitizer.sanitizerUtils.getTextDigest(transform);
-
-        String resultJson =
-                (String) textDigestFunction.map(input, sanitizer.getJsonConfiguration());
-
-        assertEquals(expected, resultJson);
-    }
 
     @SneakyThrows
     @ValueSource(strings = {"GET", "POST", "PUT", "PATCH"})
@@ -503,30 +351,7 @@ class RESTApiSanitizerImplTest {
         assertFalse(strict.contains("historyId"));
     }
 
-    @SneakyThrows
-    @ValueSource(strings = {"alice@worklytics.co, bob@worklytics.co",
-            "\"Alice Example\" <alice@worklytics.co>, \"Bob Example\" <bob@worklytics.co>",
-            "Alice.Example@worklytics.co,Bob@worklytics.co",
-    // TODO: per RFC 2822, the following SHOULD work ... but indeed lib we're using fails on it
-    // "Alice.Example@worklytics.co, , Bob@worklytics.co"
-    })
-    @ParameterizedTest
-    void pseudonymize_multivalueEmailHeaders(String headerValue) {
 
-        MapFunction mapFunction =
-                sanitizer.getPseudonymizeEmailHeaderToJson(Transform.PseudonymizeEmailHeader
-                        .builder().encoding(PseudonymEncoder.Implementations.JSON).build());
-
-        String asJson = (String) mapFunction.map(headerValue, sanitizer.getJsonConfiguration());
-
-        List<PseudonymizedIdentity> pseudonyms = sanitizer.sanitizerUtils
-                .pseudonymizeEmailHeader(sanitizer.getPseudonymizer(), headerValue);
-
-
-        assertEquals(2, pseudonyms.size());
-        assertTrue(
-                pseudonyms.stream().allMatch(p -> Objects.equals("worklytics.co", p.getDomain())));
-    }
 
 
     @CsvSource(value = {
